+++ src/components/ProfileSetup/SetLocation/KakaoMap/KakaoMapStyles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  width: 100%;
+  height: 100%;
+`;
